Work in progress
This repo/branch was superceded by the citydata repo which achieves the same thing (geonode 2.10rc4 installation, and customisation into CityData-like app).

This commit completes installation but product still seems to be a mix of Geonode 2.4 and Geonode 2.10rc4.

The latest error was internal server error 500 on / with detail about missing monitoring tables in postgresl database. That means an older version of the database.